requirements.txt
web3
python-dotenv
setuptools
langchain-openai